+++ gazebo/sensors/Sensor.cc
@@ -396,9 +396,9 @@ void Sensor::FillMsg(msgs::Sensor &_msg)
     camMsg->set_horizontal_fov(cam->HFOV().Radian());
     camMsg->mutable_image_size()->set_x(camSensor->GetImageWidth());
     camMsg->mutable_image_size()->set_y(camSensor->GetImageHeight());
-    camMsg->set_image_format(cam->GetImageFormat());
-    camMsg->set_near_clip(cam->GetNearClip());
-    camMsg->set_far_clip(cam->GetFarClip());
+    camMsg->set_image_format(cam->ImageFormat());
+    camMsg->set_near_clip(cam->NearClip());
+    camMsg->set_far_clip(cam->FarClip());
     auto distortion = cam->GetDistortion();
     if (distortion)
     {
@@ -449,4 +449,3 @@ void Sensor::DisconnectUpdated(event::ConnectionPtr &_c)
 {
   this->updated.Disconnect(_c);
 }
-